+++ pyproject.toml
@@ -59,6 +59,8 @@ black = "^24.3.0"
 pytest-asyncio = "^0.21.0"
 pytest = "^7.3.2"
 pytest-cov = "^4.1.0"
+pytest-md = "^0.2.0"
+pytest-httpserver = "^1.0.10"
 mypy = "^1.5.1"
 types-protobuf = "^4.24.0.2"
 pre-commit = "^3.2.2"
@@ -72,8 +74,6 @@ types-python-jose = "^3.3.4.8"
 openai = "^1.12.0"
 json5 = "^0.9.17"
 asgi-lifespan = "^2.1.0"
-pytest-md = "^0.2.0"
-pytest-httpserver = "^1.0.10"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -139,7 +139,6 @@ addopts = "--durations=1 -s --cov=lnbits --cov-report=xml"
 testpaths = [
   "tests/api",
   "tests/unit",
-  "tests/regtest"
 ]
 
 [tool.black]
